+++ Source/Core/Duality/Input/Backend/GlobalGamepadInputSource.cs
@@ -69,13 +69,13 @@ namespace Duality.Backend.DefaultOpenTK
 			{
 				switch (axis)
 				{
-					case GamepadAxis.LeftTrigger:		return MathF.Clamp(this.state.Triggers.Left, 0.0f, 1.0f);
-					case GamepadAxis.LeftThumbstickX:	return MathF.Clamp(this.state.ThumbSticks.Left.X, -1.0f, 1.0f);
-					case GamepadAxis.LeftThumbstickY:	return MathF.Clamp(this.state.ThumbSticks.Left.Y, -1.0f, 1.0f);
-
-					case GamepadAxis.RightTrigger:		return MathF.Clamp(this.state.Triggers.Right, 0.0f, 1.0f);
-					case GamepadAxis.RightThumbstickX:	return MathF.Clamp(this.state.ThumbSticks.Right.X, -1.0f, 1.0f);
-					case GamepadAxis.RightThumbstickY:	return MathF.Clamp(this.state.ThumbSticks.Right.Y, -1.0f, 1.0f);
+					case GamepadAxis.LeftTrigger:		return (float)MathF.Clamp(this.state.Triggers.Left, 0.0f, 1.0f);
+					case GamepadAxis.LeftThumbstickX:	return (float)MathF.Clamp(this.state.ThumbSticks.Left.X, -1.0f, 1.0f);
+					case GamepadAxis.LeftThumbstickY:	return (float)MathF.Clamp(this.state.ThumbSticks.Left.Y, -1.0f, 1.0f);
+															   
+					case GamepadAxis.RightTrigger:		return (float)MathF.Clamp(this.state.Triggers.Right, 0.0f, 1.0f);
+					case GamepadAxis.RightThumbstickX:	return (float)MathF.Clamp(this.state.ThumbSticks.Right.X, -1.0f, 1.0f);
+					case GamepadAxis.RightThumbstickY:	return (float)MathF.Clamp(this.state.ThumbSticks.Right.Y, -1.0f, 1.0f);
 
 					default: return 0.0f;
 				}
